pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-rag-system"
version = "0.1.0"
description = "Agentic RAG system for legal document retrieval and question answering"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "llama-index>=0.11.0",
    "llama-index-vector-stores-qdrant>=0.3.0",
    "llama-index-embeddings-huggingface>=0.3.0",
    "qdrant-client>=1.12.0",
    "neo4j>=5.25.0",
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
